requirements.txt
Pillow
openpyxl
pandas
pytest

# Note: tkinter may require OS-level installation.
